fix(ingestion): separate tokenizer integration tests, capture tokenize errors, and filter empty-tag examples
